backend/requirements.txt
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
Pillow==10.1.0
drf-spectacular==0.26.5
python-decouple==3.4
django-cors-headers==4.3.0
channels[daphne]==4.0.0
django-redis==5.4.0
requests==2.31.0
channels-redis==4.1.0
azure-ai-contentsafety==1.0.0